ike-init: Switch to an alternative config if proposals don't match
This way we don't rely on the order of equally matching configs anymore
(which is actually tricky in vici) and this also doesn't require
repeating weak algorithms in all configs that might potentially be selected
if there are some clients that require them.

There is currently no ordering, so an explicitly configured exactly matching
proposal isn't a better match than e.g. the default proposal that also
contains the proposed algorithms.