wip: settings: Don't allow dots in section/key names anymore
wip: The next release is a major-ish version, so we might consider making
this breaking change that will often force people to define their logfile
paths in the respective filelog sub-section. But it avoids problems created
by dots in names when looking up values (in particular if doing so
recursively).
The only problem is that the parsing error might go unnoticed after
updating to the new release (strongswan.conf will not be loaded, but
charon and swanctl run fine without it).
We could possibly add some kind of compatibility handling (e.g. allow
dots in sections called filelog - but that's not ideal), or we don't
allow executables to continue if the file exists but is invalid.
Another option is to deprecate using dots in section names (in
particular for log file paths if they are defined as section names)
and warn about this, and e.g. recommend using the `path` option, for a few
versions before we remove support for dots in section names eventually.