testing: Print complete command output if test fails in verbose mode
This could produce quite a lot of output if something is not found in a
log file, which is why this is only done in verbose mode, but is quite helpful
to debug why a pattern didn't match.

We only get stdout from SSH, so the stderr redirection is only really
for errors ssh itself produces.