dhcp: Rebind to client port if server port is in use
If another process binds to the server DHCP port on a specific address
we can't bind to 0.0.0.0. Try binding the client port before failing.
That the port is already bound could indicate that no ICMP port
unreachables will be generated later, so it might be fine to just use
the client port.